refactor: extract resume tailor helpers to resumeTailorHelpers.js
